Don't rely on cut to handle voices with multiple words
Using `say -v ? | cut -f1 -d' '` to get the list of voices cuts off
everything after the first word in voices with multiple words in them.
This is fine in the case where the first word is unambiguous as the say
command handles the shortened form.  However, we should not rely on this
being unambiguous.